Implemented a modular test api.
