+++ src/main.rs
@@ -1,10 +1,7 @@
 //! main.rs
 
-use sqlx::postgres::PgPoolOptions;
-use std::net::TcpListener;
 use zero2prod::configuration::get_configuration;
-use zero2prod::email_client::EmailClient;
-use zero2prod::startup::run;
+use zero2prod::startup::Application;
 use zero2prod::telemetry::{get_subscriber, init_subscriber};
 
 #[tokio::main]
@@ -14,23 +11,7 @@ async fn main() -> Result<(), std::io::Error> {
 
     // Panic if we can't read configuration
     let configuration = get_configuration().expect("Failed to read configuration.");
-    let connection_pool = PgPoolOptions::new().connect_lazy_with(configuration.database.with_db());
-    let sender_email = configuration
-        .emailclient
-        .sender()
-        .expect("Invalid sender email address.");
-    let timeout = configuration.emailclient.timeout();
-    let email_client = EmailClient::new(
-        configuration.emailclient.base_url,
-        sender_email,
-        configuration.emailclient.token,
-        timeout,
-    );
-    let address = format!(
-        "{}:{}",
-        configuration.application.host, configuration.application.port
-    );
-    let listener = TcpListener::bind(address)?;
-    run(listener, connection_pool, email_client)?.await?;
+    let application = Application::build(configuration).await?;
+    application.run_until_stopped().await?;
     Ok(())
 }
