+++ src/startup.rs
@@ -1,13 +1,57 @@
 //! src/startup.rs
 
+use crate::configuration::{DatabaseSettings, Settings};
 use crate::email_client::EmailClient;
 use crate::routes::{health_check, subscribe};
 use actix_web::{dev::Server, web, web::Data, App, HttpServer};
-use sqlx::PgPool;
+use sqlx::{postgres::PgPoolOptions, PgPool};
 use std::net::TcpListener;
 use tracing_actix_web::TracingLogger;
 
-pub fn run(
+pub struct Application {
+    port: u16,
+    server: Server,
+}
+
+impl Application {
+    pub async fn build(configuration: Settings) -> Result<Self, std::io::Error> {
+        let connection_pool = get_connection_pool(&configuration.database);
+        let sender_email = configuration
+            .emailclient
+            .sender()
+            .expect("Invalid sender email address.");
+        let timeout = configuration.emailclient.timeout();
+        let email_client = EmailClient::new(
+            configuration.emailclient.base_url,
+            sender_email,
+            configuration.emailclient.token,
+            timeout,
+        );
+        let address = format!(
+            "{}:{}",
+            configuration.application.host, configuration.application.port
+        );
+        let listener = TcpListener::bind(address)?;
+        let port = listener.local_addr().unwrap().port();
+        let server = run(listener, connection_pool, email_client)?;
+
+        Ok(Self { port, server })
+    }
+
+    pub fn port(&self) -> u16 {
+        self.port
+    }
+
+    pub async fn run_until_stopped(self) -> Result<(), std::io::Error> {
+        self.server.await
+    }
+}
+
+pub fn get_connection_pool(configuration: &DatabaseSettings) -> PgPool {
+    PgPoolOptions::new().connect_lazy_with(configuration.with_db())
+}
+
+fn run(
     listener: TcpListener,
     db_pool: PgPool,
     email_client: EmailClient,
