+++ tests/api/main.rs
@@ -1,0 +1,5 @@
+//! tests/api/main.rs
+
+mod health_check;
+mod helpers;
+mod subscriptions;
