--- tests/health_check.rs
@@ -1,205 +1,0 @@
-//! tests/health_check.rs
-
-use once_cell::sync::Lazy;
-use sqlx::{Connection, Executor, PgConnection, PgPool};
-use std::net::TcpListener;
-use uuid::Uuid;
-use zero2prod::configuration::{get_configuration, DatabaseSettings};
-use zero2prod::email_client::EmailClient;
-use zero2prod::startup::run;
-use zero2prod::telemetry::{get_subscriber, init_subscriber};
-
-static TRACING: Lazy<()> = Lazy::new(|| {
-    let default_filter_level = "info".to_string();
-    let subscriber_name = "test".to_string();
-    // We cannot assign the output of `get_subscriber` to a variable based on the
-    // value TEST_LOG` because the sink is part of the type returned by
-    // `get_subscriber`, therefore they are not the same type. We could work around
-    // it, but this is the most straight-forward way of moving forward.
-    if std::env::var("TEST_LOG").is_ok() {
-        let subscriber = get_subscriber(subscriber_name, default_filter_level, std::io::stdout);
-        init_subscriber(subscriber);
-    } else {
-        let subscriber = get_subscriber(subscriber_name, default_filter_level, std::io::sink);
-        init_subscriber(subscriber);
-    }
-});
-
-pub struct TestApp {
-    pub address: String,
-    pub db_pool: PgPool,
-}
-
-/// Spin up an instance of our application
-/// and returns its address (i.e. http://localhost:XXXX)
-async fn spawn_app() -> TestApp {
-    // The first time `initialize` is invoked the code in `TRACING` is executed.
-    // All other invocations will instead skip execution.
-    Lazy::force(&TRACING);
-
-    let listener = TcpListener::bind("localhost:0").expect("Failed to bind random port");
-    // We retrieve the port assigned to us by the OS
-    let port = listener.local_addr().unwrap().port();
-    let address = format!("http://localhost:{}", port);
-
-    let mut configuration = get_configuration().expect("Failed to read configuration.");
-    configuration.database.database_name = Uuid::new_v4().to_string();
-    let connection_pool = configure_database(&configuration.database).await;
-
-    let sender_email = configuration
-        .emailclient
-        .sender()
-        .expect("Invalid sender email address.");
-    let timeout = configuration.emailclient.timeout();
-    let email_client = EmailClient::new(
-        configuration.emailclient.base_url,
-        sender_email,
-        configuration.emailclient.token,
-        timeout,
-    );
-
-    let server =
-        run(listener, connection_pool.clone(), email_client).expect("Failed to bind address");
-    let _ = tokio::spawn(server);
-
-    TestApp {
-        address,
-        db_pool: connection_pool,
-    }
-}
-
-pub async fn configure_database(config: &DatabaseSettings) -> PgPool {
-    // Create database
-    let mut connection = PgConnection::connect_with(&config.without_db())
-        .await
-        .expect("Failed to connect to Postgres");
-    connection
-        .execute(format!(r#"CREATE DATABASE "{}";"#, config.database_name).as_str())
-        .await
-        .expect("Failed to create database.");
-
-    // Migrate database
-    let connection_pool = PgPool::connect_with(config.with_db())
-        .await
-        .expect("Failed to connect to Psotgres.");
-    sqlx::migrate!("./migrations")
-        .run(&connection_pool)
-        .await
-        .expect("Failed to migrate the database.");
-
-    connection_pool
-}
-
-// `tokio::test` is the testing equivalent of `tokio::main`.
-// It also spares you from having to specify the `#[test]` attribute.
-//
-// You can inspect what code gets generated using
-// `cargo expand --test health_check` (<- name of the test file)
-#[tokio::test]
-async fn health_check_works() {
-    // Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-
-    // Act
-    let response = client
-        .get(&format!("{}/health_check", &test_app.address))
-        .send()
-        .await
-        .expect("Failed to execute request.");
-
-    // Assert
-    assert!(response.status().is_success());
-    assert_eq!(Some(0), response.content_length());
-}
-
-#[tokio::test]
-async fn subscribe_returns_a_200_for_valid_form_data() {
-    // Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-
-    // Act
-    let body = "name=le%20guin&email=ursula_le_guin%40gmail.com";
-    let response = client
-        .post(&format!("{}/subscriptions", &test_app.address))
-        .header("Content-Type", "application/x-www-form-urlencoded")
-        .body(body)
-        .send()
-        .await
-        .expect("Failed to execute request.");
-
-    // Assert
-    assert_eq!(200, response.status().as_u16());
-
-    let saved = sqlx::query!("SELECT email, name FROM subscriptions")
-        .fetch_one(&test_app.db_pool)
-        .await
-        .expect("Failed to fetch saved subscription.");
-
-    assert_eq!(saved.email, "ursula_le_guin@gmail.com");
-    assert_eq!(saved.name, "le guin");
-}
-
-#[tokio::test]
-async fn subscribe_returns_a_400_when_fields_are_present_but_invalid() {
-    // Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-    let test_cases = vec![
-        ("name=&email=ursula_le_guin%40gmail.com", "empty name"),
-        ("name=Ursula&email=", "empty email"),
-        ("name=Ursula&email=definitely-not-an-email", "invalid email"),
-    ];
-
-    for (body, description) in test_cases {
-        // Act
-        let response = client
-            .post(&format!("{}/subscriptions", &test_app.address))
-            .header("Content-Type", "application/x-www-form-urlencoded")
-            .body(body)
-            .send()
-            .await
-            .expect("Failed to execute request.");
-
-        // Assert
-        assert_eq!(
-            400,
-            response.status().as_u16(),
-            "The API did not return a 400 Bad Request when the payload was {}.",
-            description
-        );
-    }
-}
-
-#[tokio::test]
-async fn subscribe_returns_a_400_when_data_is_missing() {
-    // Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-    let test_cases = vec![
-        ("name=le%20guin", "missing the email"),
-        ("email=ursula_le_guin%40gmail.com", "missing the name"),
-        ("", "missing both name and email"),
-    ];
-
-    for (invalid_body, error_message) in test_cases {
-        // Act
-        let response = client
-            .post(&format!("{}/subscriptions", &test_app.address))
-            .header("Content-Type", "application/x-www-form-urlencoded")
-            .body(invalid_body)
-            .send()
-            .await
-            .expect("Failed to execute request.");
-
-        // Assert
-        assert_eq!(
-            400,
-            response.status().as_u16(),
-            // Additional customized error message on test failure
-            "The API did not fail with 400 Bas Request when payload was {}.",
-            error_message
-        );
-    }
-}
